requirements/requirements.txt
PyYAML==3.13
stringcase==1.2.0